FDT: Preview Language Preference
The FDT selected language will now affect the preview pane. This is done
by switching out the appropriate files containing the specific language
for the preview content. The FDT does not translate, so this was the
best alternative to the time intensive development of a message based
system for all languages for the preview content similar to that done
already with the text strings on the rest of the FDT.